+++ docs/tutorials/typescript/digital-ocean-deployment/index.mdx
@@ -1,0 +1,50 @@
+---
+id: index
+title: Deploy to Digital Ocean Droplet
+description: Deploy the WorkQueue Slackbot to production.
+sidebar_label: Deploy to Production
+---
+
+## Create a droplet
+
+- Add SSH user to droplet
+  - possible helpful links:
+    - https://docs.digitalocean.com/products/droplets/how-to/add-ssh-keys/
+      - https://superuser.com/questions/772660/howto-force-ssh-to-use-a-specific-private-key
+        for me that is `ssh -i ~/.ssh/id_rsa_digitalocean`
+- Add public key to file on server:
+  To add your public key to the **`authorized_keys`** file on the server:
+  1. Display your public key with **`cat ~/.ssh/id_rsa_digitalocean.pub`** and copy the output.
+  2. Log into your DigitalOcean account, go to the Droplets section, and access your Droplet's console.
+  3. As **`root`**, open the **`~/.ssh/authorized_keys`** file with a text editor like **`nano`** or **`vi`**.
+  4. Paste your public key at the end of this file, save it, and then exit the editor.
+     If you have an existing key with a special name you can specify it when making the call
+     `ssh -i <private key filename>`
+- Now you can ssh into server
+
+Register your domain and point it at the IP
+
+setup ngnix - more or less follow this tutorial (Steps 1 & 2) https://www.digitalocean.com/community/tutorials/how-to-configure-nginx-as-a-reverse-proxy-on-ubuntu-22-04
+
+Make sure to include the full server block set up here?
+https://www.digitalocean.com/community/tutorials/how-to-install-nginx-on-ubuntu-20-04#step-5-setting-up-server-blocks-recommended
+
+Not sure if the /var/www/your-domain/html/index.html is really needed tho…
+
+Use cert bot to add cert:
+
+https://www.digitalocean.com/community/tutorials/how-to-secure-nginx-with-let-s-encrypt-on-ubuntu-20-04
+
+Generate a separate Cert for Cloud
+
+install tcld → https://docs.temporal.io/cloud/tcld#install-tcld
+
+generate using tcld → https://docs.temporal.io/cloud/certificates#use-tcld-to-generate-certificates
+
+Add certs to a env file
+
+Clone the repo on the droplet
+
+Install pm2
+
+Run the slack bot and temporal app with pm2
