+++ docs/tutorials/typescript/work-queue-slack-app/code/bot/README.md
@@ -1,0 +1,1 @@
+# iq-slack-bot
