+++ docs/tutorials/typescript/work-queue-slack-app/code/bot/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "iq-slack-bot",
+  "version": "1.0.0",
+  "description": "Temporal IQ Slack bot",
+  "main": "slack_bot.ts",
+  "repository": "git@github.com:temporalio/temporal-iq.git",
+  "author": "Cully Wakelin <cully@temporal.io>",
+  "license": "MIT",
+  "private": true,
+  "scripts": {
+    "start": "ts-node slack_bot.ts"
+  },
+  "dependencies": {
+    "@slack/bolt": "^3.19.0",
+    "@slack/web-api": "^6.12.0",
+    "@temporalio/client": "^1.7.4",
+    "crypto": "^1.0.1",
+    "date-fns": "^3.6.0",
+    "dotenv": "^16.0.3"
+  },
+  "devDependencies": {
+    "@types/node": "^20.14.7",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.5"
+  }
+}
