+++ docs/tutorials/typescript/work-queue-slack-app/code/bot/tsconfig.json
@@ -1,0 +1,21 @@
+{
+    "compilerOptions": {
+      "target": "ES2020",                // Specify ECMAScript target version
+      "module": "commonjs",              // Specify module code generation
+      "strict": true,                    // Enable all strict type-checking options
+      "esModuleInterop": true,           // Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports
+      "skipLibCheck": true,              // Skip type checking of declaration files
+      "forceConsistentCasingInFileNames": true,  // Disallow inconsistently-cased references to the same file
+      "outDir": "./dist",                // Redirect output structure to the directory
+      "rootDir": "./src",                // Specify the root directory of input files
+      "resolveJsonModule": true,         // Include modules imported with .json extension
+      "allowSyntheticDefaultImports": true,  // Allow default imports from modules with no default export
+      "moduleResolution": "node",        // Use Node's module resolution algorithm
+      "isolatedModules": true,           // Ensure TypeScript files can be safely transpiled using transpileModule or babel
+      "types": ["node"],                 // Specify type declarations
+      "lib": ["ES2020", "DOM"]           // Specify library files to be included in the compilation
+    },
+    "include": ["src/**/*"],              // Include all .ts files in the src directory
+    "exclude": ["node_modules", "dist"]   // Exclude node_modules and dist directories
+}
+
