+++ docs/tutorials/typescript/work-queue-slack-app/code/temporal-application/jest.config.js
@@ -1,0 +1,20 @@
+// module.exports = {
+//   preset: "ts-jest",
+//   testEnvironment: "node",
+//   testPathIgnorePatterns: ["/node_modules/", "/dist/"],
+//   globals: {
+//     "ts-jest": {
+//       tsconfig: "tsconfig.json",
+//     },
+//   },
+// };
+
+/** @type {import('ts-jest').JestConfigWithTsJest} */
+module.exports = {
+  preset: "ts-jest",
+  testEnvironment: "node",
+  moduleFileExtensions: ["ts", "js"],
+  transform: {
+    "^.+\\.ts$": "ts-jest",
+  },
+};
