+++ docs/tutorials/typescript/work-queue-slack-app/code/temporal-application/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "iq-temporal-application",
+  "version": "1.0.0",
+  "description": "",
+  "main": "./src/worker.ts",
+  "scripts": {
+    "start": "ts-node ./src/worker.ts",
+    "test": "jest"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/temporalio/temporal-learning.git"
+  },
+  "author": "cully@temporal.io",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/temporalio/temporal-learning/issues"
+  },
+  "dependencies": {
+    "@temporalio/worker": "^1.10.1",
+    "dotenv": "^16.1.3",
+    "path": "^0.12.7"
+  },
+  "devDependencies": {
+    "@temporalio/client": "^1.10.1",
+    "@temporalio/nyc-test-coverage": "^1.10.1",
+    "@temporalio/testing": "^1.10.1",
+    "@types/node": "^20.14.2",
+    "ts-jest": "^29.1.4",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.5",
+    "uuid": "^10.0.0"
+  }
+}
