+++ docs/tutorials/typescript/work-queue-slack-app/code/temporal-application/src/workflows/index.ts
@@ -1,0 +1,1 @@
+export { workqueue } from "./workqueue";
