--- docs/tutorials/typescript/workqueue-slackbot/code/bot/README.md
@@ -1,1 +1,0 @@
-# iq-slack-bot
