--- docs/tutorials/typescript/workqueue-slackbot/code/temporal-application/package.json
@@ -1,50 +1,0 @@
-{
-  "name": "iq-temporal-application",
-  "version": "1.0.0",
-  "description": "",
-  "main": "./src/worker.ts",
-  "scripts": {
-    "start": "ts-node ./src/worker.ts",
-    "test": "jest"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/temporalio/temporal-iq.git"
-  },
-  "author": "cully@temporal.io",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/temporalio/temporal-iq/issues"
-  },
-  "homepage": "https://github.com/temporalio/temporal-iq#readme",
-  "dependencies": {
-    "@octokit/rest": "^19.0.11",
-    "@slack/web-api": "^6.8.1",
-    "@temporalio/activity": "^1.10.1",
-    "@temporalio/worker": "^1.10.1",
-    "anzip": "^0.2.0",
-    "arraybuffer-to-buffer": "^0.0.7",
-    "axios": "^1.4.0",
-    "dotenv": "^16.1.3",
-    "fs-extra": "^11.1.1",
-    "gray-matter": "^4.0.3",
-    "module": "^1.2.5",
-    "mongodb": "^5.5.0",
-    "node-fetch": "^3.3.1",
-    "path": "^0.12.7",
-    "readdirp": "^3.6.0"
-  },
-  "devDependencies": {
-    "@temporalio/client": "^1.10.1",
-    "@temporalio/nyc-test-coverage": "^1.10.1",
-    "@temporalio/testing": "^1.10.1",
-    "@types/jest": "^29.5.12",
-    "@types/node": "^20.14.2",
-    "@types/uuid": "^9.0.8",
-    "jest": "^29.7.0",
-    "ts-jest": "^29.1.4",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.4.5",
-    "uuid": "^10.0.0"
-  }
-}
