--- docs/tutorials/typescript/workqueue-slackbot/code/temporal-application/src/workflows/index.ts
@@ -1,1 +1,0 @@
-export { workqueue } from "./workqueue";
