--- docs/tutorials/typescript/workqueue-slackbot/index.mdx
@@ -1,44 +1,0 @@
----
-id: index
-title: Build a Work Queue Slackbot with Temporal in Typescript
-description: Learn Temporal and its features while building a Work Queue Slackbot application
-sidebar_label: Build a Work Queue Slackbot with Temporal in Typescript
----
-
-Competency
-
-By the end of this tutorial, participants will be proficient in implementing an "Entity" pattern Temporal Workflow to manage state without a traditional database, specifically within the context of a Slack bot use case. They will be able to handle signals, queries, cancellation, and the continue-as-new feature using TypeScript.
-
-Assessment of Learning
-
-Participants' understanding will be assessed through a combination of practical exercises and a final project. The practical exercises will include tasks such as setting up Temporal Workflows, handling different types of signals, and implementing state management. The final project will require participants to create a fully functional Slack bot that uses Temporal Workflows to manage state across multiple channels.
-
-Learning Objectives
-Remembering: Recall the basic concepts and components of Temporal Workflows.
-
-Define key terms such as signals, queries, cancellation, and continue-as-new.
-Identify the core components of a Temporal Workflow in TypeScript.
-
-Understanding: Explain how Temporal Workflows can be used to manage state in a Slack bot.
-
-Describe the advantages of using Temporal Workflows over traditional database approaches for state management.
-Summarize the process of setting up a Temporal Workflow environment.
-
-Applying: Implement a basic Temporal Workflow for managing state in a Slack bot.
-
-Write TypeScript code to set up a Temporal Workflow.
-Demonstrate how to handle signals and queries within a Temporal Workflow.
-Analyzing: Examine and debug a Temporal Workflow implementation for a Slack bot.
-
-Analyze common issues that may arise when using Temporal Workflows for state management.
-Compare different approaches to handling state in a Slack bot using Temporal Workflows.
-
-Evaluating: Assess the effectiveness of a Temporal Workflow implementation in a Slack bot.
-
-Evaluate the performance and scalability of Temporal Workflows in managing state.
-Critique the implementation and suggest improvements.
-
-Creating: Design and develop a comprehensive Temporal Workflow for a Slack bot that includes advanced features.
-
-Create a fully functional Slack bot that uses Temporal Workflows to manage state across multiple channels.
-Integrate advanced features such as cancellation and continue-as-new into the Temporal Workflow.
