+++ snipsync.config.yaml
@@ -1,79 +1,84 @@
 origins:
-  - owner: temporalio
-    repo: money-transfer-project-template-dotnet
-  - owner: temporalio
-    repo: money-transfer-project-template-python
-  - owner: temporalio
-    repo: money-transfer-project-template-java
-    ref: 4e1ca12d8d1c058c67d810efe8dcf68f7b318208
-  - owner: temporalio
-    repo: hello-world-project-template-java
-  - owner: temporalio
-    repo: money-transfer-project-template-ts
-  - owner: temporalio
-    repo: money-transfer-project-template-go
-  - owner: temporalio
-    repo: hello-world-project-template-go
-  - owner: temporalio
-    repo: subscription-workflow-project-template-go
-  - owner: temporalio
-    repo: subscription-workflow-project-template-java
-  - owner: temporalio
-    repo: subscription-workflow-project-template-typescript
-  - owner: temporalio
-    repo: subscription-workflow-project-template-php
-  - owner: temporalio
-    repo: samples-php
-    ref: master
-  - owner: temporalio
-    repo: hello-world-project-template-python
-  - owner: temporalio
-    repo: samples-typescript
-  # Some samples were left in the SDK code itself because it has better testing
-  # capabilities than the samples repo
-  - owner: temporalio
-    repo: temporal-ecommerce
-  - owner: temporalio
-    repo: background-checks
-  - owner: temporalio
-    repo: email-subscription-project-go
-  - owner: temporalio
-    repo: data-pipeline-project-python
-  - owner: temporalio
-    repo: email-subscription-project-python
-  - owner: temporalio
-    repo: money-transfer-project-java
-    ref: cloud
-  - owner: temporalio
-    repo: money-transfer-project-template-python
-    ref: cloud
-  - owner: temporalio
-    repo: money-transfer-project-template-ts
-    ref: cloud
-  - owner: temporalio
-    repo: money-transfer-project-template-go
-    ref: cloud
-  # Background Check tutorial code in the learn repo
+  # - owner: temporalio
+  #   repo: money-transfer-project-template-dotnet
+  # - owner: temporalio
+  #   repo: money-transfer-project-template-python
+  # - owner: temporalio
+  #   repo: money-transfer-project-template-java
+  #   ref: 4e1ca12d8d1c058c67d810efe8dcf68f7b318208
+  # - owner: temporalio
+  #   repo: hello-world-project-template-java
+  # - owner: temporalio
+  #   repo: money-transfer-project-template-ts
+  # - owner: temporalio
+  #   repo: money-transfer-project-template-go
+  # - owner: temporalio
+  #   repo: hello-world-project-template-go
+  # - owner: temporalio
+  #   repo: subscription-workflow-project-template-go
+  # - owner: temporalio
+  #   repo: subscription-workflow-project-template-java
+  # - owner: temporalio
+  #   repo: subscription-workflow-project-template-typescript
+  # - owner: temporalio
+  #   repo: subscription-workflow-project-template-php
+  # - owner: temporalio
+  #   repo: samples-php
+  #   ref: master
+  # - owner: temporalio
+  #   repo: hello-world-project-template-python
+  # - owner: temporalio
+  #   repo: samples-typescript
+  # # Some samples were left in the SDK code itself because it has better testing
+  # # capabilities than the samples repo
+  # - owner: temporalio
+  #   repo: temporal-ecommerce
+  # - owner: temporalio
+  #   repo: background-checks
+  # - owner: temporalio
+  #   repo: email-subscription-project-go
+  # - owner: temporalio
+  #   repo: data-pipeline-project-python
+  # - owner: temporalio
+  #   repo: email-subscription-project-python
+  # - owner: temporalio
+  #   repo: money-transfer-project-java
+  #   ref: cloud
+  # - owner: temporalio
+  #   repo: money-transfer-project-template-python
+  #   ref: cloud
+  # - owner: temporalio
+  #   repo: money-transfer-project-template-ts
+  #   ref: cloud
+  # - owner: temporalio
+  #   repo: money-transfer-project-template-go
+  #   ref: cloud
+  # # Background Check tutorial code in the learn repo
+  # - files:
+  #     pattern: "./docs/tutorials/go/background-check/code/**/*.go"
+  #     owner: "temporalio"
+  #     repo: "temporal-learning"
+  #     ref: "main"
+  # - files:
+  #     pattern: "./docs/tutorials/java/background-check/code/**/*.java"
+  #     owner: "temporalio"
+  #     repo: "temporal-learning"
+  #     ref: "main"
+  # - files:
+  #     pattern: "./docs/tutorials/python/background-check/code/**/*.py"
+  #     owner: "temporalio"
+  #     repo: "temporal-learning"
+  #     ref: "main"
+  # - files:
+  #     pattern: "./docs/tutorials/typescript/background-check/code/**/*.ts"
+  #     owner: "temporalio"
+  #     repo: "temporal-learning"
+  #     ref: "main"
   - files:
-      pattern: './docs/tutorials/go/background-check/code/**/*.go'
-      owner: 'temporalio'
-      repo: 'temporal-learning'
-      ref: 'main'
-  - files:
-      pattern: './docs/tutorials/java/background-check/code/**/*.java'
-      owner: 'temporalio'
-      repo: 'temporal-learning'
-      ref: 'main'
-  - files:
-      pattern: './docs/tutorials/python/background-check/code/**/*.py'
-      owner: 'temporalio'
-      repo: 'temporal-learning'
-      ref: 'main'
-  - files:
-      pattern: './docs/tutorials/typescript/background-check/code/**/*.ts'
-      owner: 'temporalio'
-      repo: 'temporal-learning'
-      ref: 'main'
+      pattern: "./docs/tutorials/typescript/work-queue-slack-app/code/**/*.ts"
+      owner: "temporalio"
+      repo: "temporal-learning"
+      ref: "edu-2636-workqueue-slackbot-tut"
 
 targets:
   - docs/tutorials
